requirements.txt
sqlalchemy
psycopg2
pytz
mysqlclient
pymysql
cryptography
nba_api
boto3
pandas
sqlalchemy_utils
fsspec
s3fs
flask
sqlmodel
Django>=4.2,<5.0
django-ninja>=1.0
django-cors-headers>=4.0
polars==1.35.1
PyMySQL==1.1.2
scikit-learn